[Fixed: 3] We can't handle any update that is not encapsulated in a <reference>, so ignore.
This is due (I think) to the fact that the elements are not encapsulated in a reference and, as such, cannot simply be deleted from their previous
position in the array and appended to the array because of the potential side effects of appearing further down the file.

This is going to have implications on the main content column, although it's not possible to to edit any of the actual structure with in the main
content column yet anyway.