[Committed: 106] Change the method of ordering blocks in columns
The original method of doing this was to alter the after statement of the block
in order for Magento to accurately process the sorted chidlren when rendering
the core/text_list block.  This proved to be error prone, especially when blocks
in the order chain were not included on the page.

We've now moved to overriding the core/text_list block to implement our own
strict ordering of the blocks when the 'block_order' data item is set
on the block during the layout process.